requirements.txt
numpy==1.25.2
opencv_python==4.8.0.76
Pillow==10.0.0
Pillow==10.0.0
pywin32==304
tensorboardX==2.6.2
tensorboardX==2.6.2.2
torch==2.0.1+cu118
torchvision==0.15.2+cu118
